requirements.txt


python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
selenium==4.1.0
webdriverdownloader==1.1.0.3
selenium-wire==3.0.6
pandas
openpyxl
uiautomator2==3.0.8
python-decouple==3.5
django-admin-relation-links
zstd
gspread
async-timeout
psycopg2
google-api-core==2.5.0
google-api-python-client==2.123.0
google-auth==2.6.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.3
google-cloud-core==2.2.2
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
djangorestframework
django-jsonform
pyzstd
ps
psutil==5.9.8
